pyproject.toml
[project]
name = "franking"
version = "0.1.0"
description = "A simple Interface for creating shipping labels using Deutsche Post Internetmarke for Invio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "brother-ql @ git+https://github.com/pklaus/brother_ql.git",
    "fastapi[standard]>=0.128.0",
    "inema>=1.0.4",
    "pycountry>=24.6.1",
    "python-dotenv>=1.2.1",
]
